Fix more warnings in samples and tests
These warnings are currently disabled because these files are not with
angle_code set to 1 in the gyp files to avoid gtest's code warnings.
We hope to either fix warnings in gtest or change our build system to
not use warnings for gtest.

BUG=angleproject:1003

Reviewed-on: https://chromium-review.googlesource.com/271412
Tested-by: Corentin Wallez <cwallez@chromium.org>

+++ util/linux/Linux_path_utils.cpp
@@ -19,7 +19,7 @@ std::string GetExecutablePath()
     char path[4096];
 
     ssize_t result = readlink("/proc/self/exe", path, sizeof(path) - 1);
-    if (result < 0 || result >= sizeof(path) - 1)
+    if (result < 0 || static_cast<size_t>(result) >= sizeof(path) - 1)
     {
         return "";
     }
